requirements.txt
# pip install -r requirements.txt
openai
numpy
pandas
python-dotenv
langchain
